Move efergy base entity to separate module (#126051)
